switch intptr_t to __int64; intptr_t not found by msvc6; __int64 works with msvc6; http://msdn.microsoft.com/library/default.asp?url=/library/en-us/vclib/html/_crt_standard_types.asp

+++ core/vul/vul_file_iterator.cxx
@@ -12,13 +12,15 @@
 // Declare pimpl, reset, and iteration routines
 // for each OS
 #if defined(VCL_WIN32) && !defined(__CYGWIN__)
+#include <stddef.h>
 #include <io.h>
 struct vul_file_iterator_data
 {
   vcl_string original;
   vcl_string original_dirname;
   struct _finddata_t data;
-  intptr_t handle;
+  // intptr_t handle;  // not found by msvc6
+  __int64 handle;      // works with msvc6
   vcl_string found;
   char const* name;
